refactor: extract single entity fields
